+++ erdjs/src/smartcontracts/codecs.spec.ts
@@ -1,7 +1,7 @@
 import { describe } from "mocha";
 import { assert } from "chai";
 import { Address } from "../address";
-import { AbiRegistry, StructureDefinition } from "./abi";
+import { AbiRegistry, StructureDefinition } from "./typesystem/abi";
 import { BinaryCodec, discardSuperfluousBytesInTwosComplement, discardSuperfluousZeroBytes, isMbsOne } from "./codecs";
 import { BooleanValue, NumericalValue, PrimitiveType } from "./types";
