+++ erdjs/src/smartcontracts/typesystem/interfaces.ts
@@ -1,0 +1,7 @@
+import { TypeDescriptor } from "./typeDescriptor";
+import { Type } from "./types";
+
+export interface ITypeResolver {
+    resolveTypeDescriptor(scopedTypeNames: string[]): TypeDescriptor;
+    resolveType(typeName: string): Type;
+}
